+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.4] Q1 2023
+ - systemd::start_unit: added timeout handling
+ - systemd::wait_for_system_running: added timeout handling
+ - systemd::reboot: sync journal before reboot
+ - module version: also log the short git revision
+ - explicit `cargo update` to fix RUSTSEC-2023-0034
+
 ## [0.10.3] Q1 2023
  - fixed twin update handling when key is not present at all vs. key is null